backup_exporter/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
path==15.1.2
prometheus-client==0.10.1
python-dotenv==0.17.1
requests==2.25.1
urllib3==1.26.4
